src/keyword/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.2
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gensim==4.3.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
inexactsearch==1.0.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
keybert==0.9.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
openai==1.76.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
preshed==3.0.9
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
pyspellchecker==0.8.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rake-nltk==1.0.6
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
shellingham==1.5.4
silpa_common==0.3
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundex==1.1.3
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0+cu126
torchaudio==2.7.0+cu126
torchvision==0.22.0+cu126
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
wordcloud==1.9.4
wrapt==1.17.2
